requirements.txt
streamlit
torch>=2.0
torchvision
xgboost
Pillow
numpy
opencv-python-headless